requirements.txt
Django~=4.0.5
psycopg2-binary==2.8.6
djangorestframework==3.12.4
django-cors-headers==3.10.0
pytz==2022.1
scrapingbee==1.1.7
bs4==0.0.1
matplotlib==3.5.2
